[REF] create a multi mixin
