+++ nilearn/maskers/multi_nifti_labels_masker.py
@@ -7,14 +7,14 @@ from sklearn.utils.estimator_checks import check_is_fitted
 
 from nilearn._utils.docs import fill_doc
 from nilearn._utils.niimg_conversions import iter_check_niimg
-from nilearn._utils.tags import SKLEARN_LT_1_6
+from nilearn.maskers._mixin import _MultiMixin
 from nilearn.maskers.base_masker import prepare_confounds_multimaskers
 from nilearn.maskers.nifti_labels_masker import NiftiLabelsMasker
 from nilearn.typing import NiimgLike
 
 
 @fill_doc
-class MultiNiftiLabelsMasker(NiftiLabelsMasker):
+class MultiNiftiLabelsMasker(_MultiMixin, NiftiLabelsMasker):
     """Class for extracting data from multiple Niimg-like objects \
        using labels of non-overlapping brain regions.
 
@@ -180,24 +180,6 @@ class MultiNiftiLabelsMasker(NiftiLabelsMasker):
             keep_masked_labels=keep_masked_labels,
         )
 
-    def __sklearn_tags__(self):
-        """Return estimator tags.
-
-        See the sklearn documentation for more details on tags
-        https://scikit-learn.org/1.6/developers/develop.html#estimator-tags
-        """
-        # TODO (sklearn  >= 1.6.0) remove if block
-        if SKLEARN_LT_1_6:
-            from nilearn._utils.tags import tags
-
-            return tags(masker=True, multi_masker=True)
-
-        from nilearn._utils.tags import InputTags
-
-        tags = super().__sklearn_tags__()
-        tags.input_tags = InputTags(masker=True, multi_masker=True)
-        return tags
-
     @fill_doc
     def transform_imgs(
         self, imgs_list, confounds=None, n_jobs=1, sample_mask=None
@@ -298,33 +280,4 @@ class MultiNiftiLabelsMasker(NiftiLabelsMasker):
             confounds=confounds,
             sample_mask=sample_mask,
             n_jobs=self.n_jobs,
-        )
-
-    @fill_doc
-    def fit_transform(self, imgs, y=None, confounds=None, sample_mask=None):
-        """
-        Fit to data, then transform it.
-
-        Parameters
-        ----------
-        imgs : Niimg-like object, or a :obj:`list` of Niimg-like objects
-            See :ref:`extracting_data`.
-            Data to be preprocessed
-
-        y : None
-            This parameter is unused. It is solely included for scikit-learn
-            compatibility.
-
-        %(confounds_multi)s
-
-        %(sample_mask_multi)s
-
-            .. versionadded:: 0.8.0
-
-        Returns
-        -------
-        %(signals_transform_multi_nifti)s
-        """
-        return self.fit(imgs, y=y).transform(
-            imgs, confounds=confounds, sample_mask=sample_mask
         )
